moved plotting files to plotting subdirectory
